+++ flexmeasures/api/v3_0/__init__.py
@@ -10,6 +10,7 @@ from flexmeasures.api.v3_0.users import UserAPI
 from flexmeasures.api.v3_0.assets import AssetAPI, AssetTypesAPI
 from flexmeasures.api.v3_0.health import HealthAPI
 from flexmeasures.api.v3_0.public import ServicesAPI
+from flexmeasures.api.v3_0.deprecated import SensorEntityAddressAPI
 
 
 def register_at(app: Flask):
@@ -24,3 +25,5 @@ def register_at(app: Flask):
     AssetTypesAPI.register(app, route_prefix=v3_0_api_prefix)
     HealthAPI.register(app, route_prefix=v3_0_api_prefix)
     ServicesAPI.register(app)
+
+    SensorEntityAddressAPI.register(app, route_prefix=v3_0_api_prefix)