+++ flexmeasures/cli/data_show.py
@@ -438,9 +438,10 @@ def list_data_sources(source: DataSource | None = None, show_attributes: bool = 
     required=False,
     type=str,
     default="chart-$now.png",
-    help="Format of the output file. Use dollar sign ($) to interpolate values among the following ones:"
-    " now (current time), id (id of the sensor or asset), entity_type (either 'asset' or 'sensor')"
-    " Example: 'result_file_$entity_type_$id_$now.csv' -> 'result_file_asset_1_2023-08-24T14:47:08' ",
+    help="Output filename template. You can use the following placeholders with a dollar sign ($): "
+    "$now (current timestamp), $id (sensor or asset ID), and $entity_type ('asset' or 'sensor'). "
+    "Example: 'chart_$entity_type_$id_$now.png' → 'chart_asset_1_2023-08-24T14:47:08.png'. "
+    "Note: you may need to escape the dollar sign (\\$) in your terminal.",
 )
 @click.option(
     "--resolution",
@@ -448,6 +449,15 @@ def list_data_sources(source: DataSource | None = None, show_attributes: bool = 
     type=DurationField(),
     required=False,
     help="Resolution of the data in ISO 8601 format. If not set, defaults to the minimum resolution of the sensor data. Note: Nominal durations like 'P1D' are converted to absolute timedeltas.",
+)
+@click.option(
+    "--combine-legend",
+    "combine_legend",
+    type=bool,
+    default=False,
+    required=False,
+    is_flag=True,
+    help="If True, then legend will be combined and put in the bottom of all charts, otherwise each chart for an asset can have a separate legend. Default value is False.",
 )
 def chart(
     sensors: list[Sensor] | None = None,
@@ -459,6 +469,7 @@ def chart(
     width: int | None = None,
     filename_template: str | None = None,
     resolution: timedelta | None = None,
+    combine_legend: bool = False,
 ):
     """
     Export sensor or asset charts in PNG or SVG formats. For example:
@@ -518,6 +529,7 @@ def chart(
             beliefs_before=belief_time,
             include_data=True,
             resolution=resolution,
+            combine_legend=combine_legend,
         )
 
         # remove formatType as it relies on a custom JavaScript function