+++ flexmeasures/ui/views/assets/__init__.py
@@ -1,6 +1,5 @@
 from flexmeasures.ui.views.assets.forms import AssetForm, NewAssetForm  # noqa: F401
 from flexmeasures.ui.views.assets.utils import (  # noqa: F401
-    process_internal_api_response,
     user_can_create_assets,
     user_can_delete,
     user_can_update,